complete http method test

+++ tasktracker/views.py
@@ -1,10 +1,13 @@
 from django.shortcuts import render
-from django.http import HttpResponseRedirect
+from django.http import HttpResponseRedirect, HttpResponseNotFound, HttpResponseBadRequest
 from django.shortcuts import render, reverse
 from .models import Task
 from django.db import connection
 from django.core.exceptions import ValidationError
 
+import logging
+
+logger = logging.getLogger()
 
 # View functions
 def index(request):
@@ -17,6 +20,7 @@ def index(request):
 		# renders the web page
 		return render(request, 'index.html', template_data)
 	else:
+		logger.debug(f'unauthenticated request to {request.path} from {request.user}')
 		return HttpResponseRedirect(reverse(f'login'))
 
 # Adds a new task and redirect it back to index page
@@ -33,24 +37,53 @@ def add(request):
 				task.full_clean() 
 				# if no exception was thrown, form was validated
 				# we proceed to save the task in the database
-				with connection.cursor() as cursor:
-					cursor.executescript(f"INSERT INTO tasktracker_task(user_id, status, due_date, title) VALUES ({request.user.id},'{status}', '{due_date}', '{title}')")				
+
+				# utilizes django's built-in object management to save to database
+				# this prevents SQL injection attacks
+				task.save()
+				logger.info(f'user {request.user} created new task with id {task.id}')
 			except ValidationError as e:
 				# renders the web page again with an error message
+				logger.error(f'user {request.user} attempted to create invalid task: {task}')
 				return render(request, 'add.html', {"errors": e.message_dict})
 				
 			return HttpResponseRedirect(reverse(f'tasktracker:index'))
 		else:
 			# renders the web page
+			logger.debug(f'invalid request method to {request.path} by {request.user}')
 			return render(request, 'add.html')
 	else:
+		logger.debug(f'unauthenticated request to {request.path} from {request.user}')
 		return HttpResponseRedirect(reverse(f'login'))
 
 
 # Deletes a task (based on its primary key) and redirect it back to index page
 def delete(request, pk):
-	# uses ORM to delete the task
-	task = Task.objects.get(id = pk)
-	task.delete()
-	# redirects user to index page
-	return HttpResponseRedirect(reverse(f'tasktracker:index'))
+	# requires user is authenticated
+	if request.user.is_authenticated:
+
+		# requires request is a POST so CSRF token is checked
+		if request.method == 'POST':
+			# uses ORM to delete the task
+			try:
+				task = Task.objects.get(id = pk)
+
+				# checks task belongs to request sender before deletion
+				if task.user == request.user:
+					task.delete()
+					logger.info(f'user {request.user} deleted task with id {pk}')
+				else:
+					logger.critical(f'user {request.user} attempting unauthorized delete of task with id {pk}')
+					return HttpResponseNotFound()
+			except Task.DoesNotExist as e:
+				logger.error(f'user {request.user} attempting to delete nonexistent task with id {pk}')
+				print(e)
+				return HttpResponseNotFound()
+			return HttpResponseRedirect(reverse(f'tasktracker:index'))
+			# redirects user to index page
+		else:
+			logger.debug(f'invalid request method to {request.path} by {request.user}')
+			return HttpResponseBadRequest()
+	else:
+		logger.debug(f'unauthenticated request to {request.path} from {request.user}')
+		return HttpResponseRedirect(reverse(f'login'))
